+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 #### 0.10.0 (unfinished)
 
-Public functions are now also exposed as constants, so that 
+* Public functions are now also exposed as constants, so that 
 use of 'P::someFunction' can be substituted by P::someFunction 
 without the need for quotes
+* Added `partial`, `partialRight`
 
 #### 0.9.0